* Fixed an issue where the RegionFlags and SimAccess settings were being saved incorrectly.  If your region is already having an issue on boot with parsing [region_flags] or [sim_access], then this won't solve your issue.   Default value for region_flags is 0 and default value for sim_access is 21 in the EstateSettings.XML file

+++ OpenSim/Framework/EstateSettings.cs
@@ -128,7 +128,7 @@ namespace OpenSim.Framework
             set
             {
                 m_regionFlags = value;
-                configMember.forceSetConfigurationOption("region_flags", m_regionFlags.ToString());
+                configMember.forceSetConfigurationOption("region_flags", ((uint)m_regionFlags).ToString());
             }
         }
 
@@ -141,7 +141,7 @@ namespace OpenSim.Framework
             set
             {
                 m_simAccess = value;
-                configMember.forceSetConfigurationOption("sim_access", m_simAccess.ToString());
+                configMember.forceSetConfigurationOption("sim_access", ((byte)m_simAccess).ToString());
             }
         }
